requirements.txt
asgiref==3.9.1
Django>=4.2,<4.3
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.1
